Permission Control For Author

+++ app/Http/Controllers/Author/PostController.php
@@ -106,6 +106,10 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            return redirect()->back();
+        }
+
         return view('author.post.show', compact('post'));
     }
 
@@ -117,6 +121,10 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            return redirect()->back();
+        }
+
         $categories = Category::all();
         $tags = Tag::all();
         return view('author.post.edit', compact('post', 'categories', 'tags'));
@@ -131,6 +139,10 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required',
             'image' => 'image',
@@ -194,6 +206,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if ($post->user_id != Auth::id()) {
+            return redirect()->back();
+        }
+
         if (Storage::disk('public')->exists('post/'.$post->image)){
             Storage::disk('public')->delete('post/'.$post->image);
         }
